backend/requirements.txt
# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Celery
celery[redis]==5.3.6
redis==5.0.1

# HTTP client
httpx==0.26.0
aiohttp==3.9.1

# OpenAI
openai==1.10.0

# Tree-sitter
tree-sitter==0.21.3
tree-sitter-python==0.21.0
tree-sitter-javascript==0.21.0
tree-sitter-typescript==0.21.0
tree-sitter-java==0.21.0
tree-sitter-go==0.21.0
tree-sitter-rust==0.21.0
tree-sitter-c==0.21.0
tree-sitter-cpp==0.21.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
gitpython==3.1.41

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
httpx==0.26.0
